# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FleetSaaS school bus tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles for this tenant
      responses:
        "200":
          description: Vehicle list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Add a new vehicle (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreate"
      responses:
        "201":
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    patch:
      operationId: patchVehicle
      tags: [vehicles]
      summary: Update vehicle tag (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Remove a vehicle (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FLEET SWAP ─────────────────────────────────────────────────────────────
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List all routes for this tenant
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update route name or assigned driver/vehicle (fleet swap)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdate"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /routes/{id}/stations:
    get:
      operationId: listRouteStations
      tags: [routes]
      summary: Get ordered stations for a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ordered station list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteStation"
    post:
      operationId: addRouteStation
      tags: [routes]
      summary: Add a station to a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteStationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteStation"
  /routes/{id}/stations/{routeStationId}:
    delete:
      operationId: removeRouteStation
      tags: [routes]
      summary: Remove a route-station entry by its row ID (supports duplicate stops)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: routeStationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /routes/{id}/stations/reorder:
    post:
      operationId: reorderRouteStations
      tags: [routes]
      summary: Reorder stations within a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderStationsInput"
      responses:
        "200":
          description: Reordered stations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteStation"

  # ─── GEOCODE ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Route:
      type: object
      required: [id, tenantId, name, isActive]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        driverName:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        departureTime:
          type: string
          description: "Base departure time string e.g. '06:00 AM'"
        avgSpeedKmh:
          type: integer
          description: Average fleet speed in km/h for ETA calculation
    RouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        driverId:
          type: integer
        vehicleId:
          type: integer
        departureTime:
          type: string
        avgSpeedKmh:
          type: integer
    RouteUpdate:
      type: object
      properties:
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        departureTime:
          type: string
        avgSpeedKmh:
          type: integer
    RouteStation:
      type: object
      required: [id, routeId, stationId, position]
      properties:
        id:
          type: integer
        routeId:
          type: integer
        stationId:
          type: integer
        position:
          type: integer
        direction:
          type: string
          description: "'forward' or 'return'"
        stopLabel:
          type: ["string", "null"]
          description: Optional custom label e.g. "Koteshwor (Return)"
        stationName:
          type: string
        lat:
          type: number
        lng:
          type: number
        radius:
          type: integer
        eta:
          type: ["string", "null"]
          description: "Computed ETA string e.g. '06:20 AM'"
    RouteStationInput:
      type: object
      required: [stationId]
      properties:
        stationId:
          type: integer
        position:
          type: integer
        direction:
          type: string
          description: "'forward' or 'return'"
        stopLabel:
          type: string
          description: Optional custom display label
    ReorderStationsInput:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items:
            type: integer
          description: Ordered list of route_station row IDs (supports duplicate stops)
...
    Vehicle:
      type: object
      required: [id, plateNumber, model]
      properties:
        id:
          type: integer
        plateNumber:
          type: string
        model:
          type: string
        capacity:
          type: integer
        isActive:
          type: boolean
        tag:
          type: string
          nullable: true
    VehicleCreate:
      type: object
      required: [plateNumber, model]
      properties:
        plateNumber:
          type: string
        model:
          type: string
        capacity:
          type: integer
        tag:
          type: string
          nullable: true
    VehicleUpdate:
      type: object
      properties:
        tag:
          type: string
          nullable: true
...